feat: add functional search component
